+++ src/main/sql/table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE product (
     pk bigserial primary key,
     name varchar(20) NOT NULL,
-    price bigint NOT NULL
+    price numeric(10,2) NOT NULL
 )
